Move db.js(for user) into dbs directory
