+++ backend/src/models/expense.js
@@ -1,9 +1,14 @@
 //소비 내역
 const mongoose = require('mongoose');
+import {utcToKst} from '../misc/timeZone';
 
 const expenseSchema = new mongoose.Schema({
     name : String,
-    user : mongoose.Schema.Types.ObjectId,
+    userID : String,
+    date : {
+        type: Date,
+        default: utcToKst(new Date())
+    },
     money : {
         type: Number,
         default: 0
